Store mixer on the context
