+++ packages/react-hooks/src/mixer/useMixer.ts
@@ -1,0 +1,86 @@
+import { MixerContext, MixerContextData } from '@webb-dapp/react-environment';
+import { useCall } from '@webb-dapp/react-hooks/useCall';
+import { LoggerService } from '@webb-tools/app-util';
+import { Mixer, MixerAssetGroup } from '@webb-tools/sdk-mixer';
+import { GroupId } from '@webb-tools/types/interfaces';
+import { useCallback, useContext, useMemo, useState } from 'react';
+// @ts-ignore
+import Worker from './mixer.worker';
+
+/**
+ * @name useMixer
+ * @description get Mixer context value
+ */
+export const useMixerProvider = (): MixerContextData => {
+  return useContext<MixerContextData>(MixerContext);
+};
+
+export const useMixerGroups = (): MixerAssetGroup[] => {
+  const mixerGroupIds = useCall<Array<GroupId>>('query.mixer.mixerGroupIds', []);
+
+  return useMemo(() => {
+    return (
+      mixerGroupIds?.map((id) => {
+        return new MixerAssetGroup(Number(id.toString()), 'EDG', 32);
+      }) ?? []
+    );
+  }, [mixerGroupIds]);
+};
+
+const logger = LoggerService.new('MixerUsage');
+
+export const useMixer = () => {
+  const mixerAssetGroup = useMixerGroups();
+
+  const [mixerResult, setMixerResult] = useState<Omit<MixerContextData, 'init'>>({
+    initialized: false,
+    loading: true,
+    mixer: null,
+    shouldDestroy: false,
+  });
+
+  const [called, setCalled] = useState(false);
+
+  const init = useCallback(async () => {
+    setCalled(true);
+
+    if (called || mixerResult.initialized) {
+      logger.info(`Generating new mixer`);
+      logger.error(`Mixer already initialize`);
+
+      return undefined;
+    }
+
+    logger.info(`Generating new mixer`);
+    setMixerResult((p) => ({
+      ...p,
+      loading: true,
+    }));
+
+    try {
+      const mixer = await Mixer.init(new Worker(), mixerAssetGroup);
+
+      logger.info(`Generated new mixer`);
+      logger.debug(`Generated Mixer `, mixer);
+
+      setMixerResult((p) => ({
+        ...p,
+        loading: false,
+        mixer,
+      }));
+    } catch (e) {
+      logger.error(`Mixer creation error`, e);
+    } finally {
+      setCalled(false);
+      setMixerResult((p) => ({
+        ...p,
+
+        loading: false,
+      }));
+    }
+  }, [mixerAssetGroup, mixerResult]);
+  return {
+    ...mixerResult,
+    init,
+  };
+};
